+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.7.2 (Unreleased)
 
 IMPROVEMENTS:
+
+- *New DataSource*: _alicloud_eips_ ([#110](https://github.com/terraform-providers/terraform-provider-alicloud/pull/110))
 - *New DataSource*: _alicloud_vswitches_ ([#109](https://github.com/terraform-providers/terraform-provider-alicloud/pull/109))
 
 BUG FIXES: